+++ CHANGES.markdown
@@ -1,3 +1,8 @@
+# Next
+
+* Cache timezone data in memory for performance. (panthomakos)
+* Find timezone rule using binary search for performance. (panthomakos)
+
 # 0.3.7
 
 * Added `Timezone::Zone#local_to_utc` function. (panthomakos)